% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{biasCorrected_Neyman}
\alias{biasCorrected_Neyman}
\title{Bias-corrected Neyman Sample Average Treatment Effect Estimator}
\usage{
biasCorrected_Neyman(Y, Z, pairs, pmat, xi)
}
\arguments{
\item{Y}{a I-length vector of outcome values}

\item{Z}{a I-length vector of treatment values}

\item{pairs}{a I x 2 dataframe containing the indices of observations
that form our set of matched pairs.}

\item{pmat}{a 2I x 2I matrix where the diagonals equal zero, and the
off-diagonal elements (i, j) contain the probability the ith observation has
Z = max{Z_i, Z_j} and the jth observation has Z = min{Z_i, Z_j}. We can create
a p-matrix using the make_pmatrix function.}

\item{xi}{a number in the range 0 to 0.5, the cutoff related to the treatment
assignment probability caliper.}
}
\value{
I x 2 dataframe
}
\description{
This function estimates the sample average treatment effect for a set of
matched pairs using the bias-corrected Neyman estimator.
}
\examples{
set.seed(12345)
X <- rnorm(1000, 0, 5)
Z <- X + rnorm(1000, 0, (1+sqrt(abs(X))))
Y <- X + Z + rnorm(1000, 0, 0.5)
pmat <- make_pmatrix(Z, X)
pairs <- nbp_Caliper(Z, X, pmat, xi = 0.1, M = 10000)
biasCorrected_Neyman(Y, Z, pairs, pmat, xi = 0.1)
}
