% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{classic_Neyman}
\alias{classic_Neyman}
\title{Classic Neyman Sample Average Treatment Effect Estimator}
\usage{
classic_Neyman(Y, Z, pairs)
}
\arguments{
\item{Y}{a I-length vector of outcome values}

\item{Z}{a I-length vector of treatment values}

\item{pairs}{a I x 2 dataframe containing the indices of observations
that form our set of matched pairs.}
}
\value{
I x 2 dataframe
}
\description{
This function estimates the sample average treatment effect for a set of
matched pairs using the classic Neyman estimator.
}
\examples{
set.seed(12345)
X <- rnorm(1000, 0, 5)
Z <- X + rnorm(1000, 0, (1+sqrt(abs(X))))
Y <- X + Z + rnorm(1000, 0, 0.5)
pmat <- make_pmatrix(Z, X)
pairs <- nbp_Caliper(Z, X, pmat, xi = 0.1, M = 10000)
classic_Neyman(Y, Z, pairs)
}
