% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{covAdj_Variance}
\alias{covAdj_Variance}
\title{Covariate-Adjusted Variance Estimation}
\usage{
covAdj_Variance(Y, Z, X, pairs, pmat, xi, Q)
}
\arguments{
\item{Y}{a I-length vector of outcome values}

\item{Z}{a I-length vector of treatment values}

\item{X}{a I x k matrix of covariate values}

\item{pairs}{a I x 2 dataframe containing the indices of observations
that form our set of matched pairs.}

\item{pmat}{a 2I x 2I matrix where the diagonals equal zero, and the
off-diagonal elements (i, j) contain the probability the ith observation has
Z = max{Z_i, Z_j} and the jth observation has Z = min{Z_i, Z_j}. We can create
a p-matrix using the make_pmatrix function.}

\item{xi}{a number in the range \link{0, 0.5}, the cutoff related to the treatment
assignment probability caliper.}

\item{Q}{an arbitrary I x L matrix, where L < I}
}
\value{
a 2I x 2I numeric matrix
}
\description{
This function calculates the covariate-adjusted conservative variance estimator
For the (classic or bias-corrected) Neyman estimator
}
\examples{
set.seed(12345)
X <- rnorm(1000, 0, 5)
Z <- X + rnorm(1000, 0, (1+sqrt(abs(X))))
Y <- X + Z + rnorm(1000, 0, 0.5)
pmat <- make_pmatrix(Z, X)
pairs <- nbp_caliper(Z, X, pmat, xi = 0.1, M = 10000)
covAdj_variance(Y, Z, X, pairs, pmat, xi = 0.1)
}
