% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{generate_data_dose}
\alias{generate_data_dose}
\title{Generate some example data}
\usage{
generate_data_dose(N)
}
\arguments{
\item{N}{number of observations to simulate}
}
\value{
an N x 7 matrix containing treatment, outcome, and covariates.
}
\description{
This function creates some example data using the data generation process described
in simulation 1 of Frazier, Heng, Zhou (2024). The dataframe contains a treatment
variable Z, outcome variable Y, and five covariate X1,...,X5.
}
\examples{
generate_data_dose(N = 100)
}
