% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{make_pmatrix}
\alias{make_pmatrix}
\title{Make a p-Matrix}
\usage{
make_pmatrix(Z, X, method = "linCDE")
}
\arguments{
\item{Z}{a I-length vector of treatment values}

\item{X}{a I x k matrix of covariate values}

\item{method}{currently takes either "linCDE" or "RFCDE" as values. Determines
which conditional density estimation method to use to estimate the generalized
propensity score. Defaults to "linCDE"}
}
\value{
a 2I x 2I numeric matrix
}
\description{
This function creates a 2I x 2I matrix where the diagonals equal zero, and the
off-diagonal elements (i, j) contain the probability the ith observation has
Z = max{Z_i, Z_j} and the jth observation has Z = min{Z_i, Z_j}.
}
\examples{
set.seed(12345)
X <- rnorm(1000, 0, 5)
Z <- X + rnorm(1000, 0, (1+sqrt(abs(X))))
make_pmatrix(Z, X)
}
