% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{nbp_Caliper}
\alias{nbp_Caliper}
\title{non-bipartite matching with treatment assignment caliper}
\usage{
nbp_Caliper(Z, X, pmat, xi, M)
}
\arguments{
\item{Z}{a I-length vector of treatment values}

\item{X}{a I x k matrix of covariate values}

\item{pmat}{a 2I x 2I matrix where the diagonals equal zero, and the
off-diagonal elements (i, j) contain the probability the ith observation has
Z = max{Z_i, Z_j} and the jth observation has Z = min{Z_i, Z_j}. We can create
a p-matrix using the make_pmatrix function.}

\item{xi}{a number in the range 0 to 0.5, the cutoff related to the treatment
assignment probability caliper.}

\item{M}{an integer determining the penalty of the treatment assignment
probability caliper. If a potential matched pair between observations i and j
has treatment assignment probability less than xi or greater than 1-xi, we add
M to the distance matrix in the (i, j) and (j, i) entry.}
}
\value{
I x 2 dataframe
}
\description{
This function creates a I x 2 dataframe containing the indices of observations
that form our set of matched pairs. It uses the nbpMatch package along with
a p-atrix in order to create I matched pairs using a treatment assignment
caliper.
}
\examples{
set.seed(12345)
X <- rnorm(1000, 0, 5)
Z <- X + rnorm(1000, 0, (1+sqrt(abs(X))))
pmat <- make_pmatrix(Z, X)
nbp_Caliper(Z, X, pmat, xi = 0.1, M = 10000)
}
